+++ Content.IntegrationTests/Tests/Commands/SuicideCommandTests.cs
@@ -363,7 +363,7 @@ public sealed class SuicideCommandTests
                     Assert.That(mobStateSystem.IsDead(player, mobStateComp));
                     Assert.That(entManager.TryGetComponent<GhostComponent>(mindComponent.CurrentEntity, out var ghostComp) &&
                                 !ghostComp.CanReturnToBody);
-                    Assert.That(damageableComp.Damage.DamageDict["Blunt"], Is.EqualTo(lethalDamageThreshold / 2)); // WD EDIT
+                    Assert.That(damageableComp.Damage.DamageDict["Slash"], Is.EqualTo(lethalDamageThreshold / 2));
                 });
         });
